PDC Project/parallel_image_processing/requirements.txt
pillow>=10.0.0
matplotlib
